+++ kite/src/test/kotlin/com/tang/kite/io/ResourcesTest.kt
@@ -1,8 +1,7 @@
 package com.tang.kite.io
 
-import org.junit.jupiter.api.Test
-
-import org.junit.jupiter.api.Assertions.*
+import kotlin.test.Test
+import kotlin.test.assertNotNull
 
 /**
  * @author Tang