+++ kite/src/test/kotlin/com/tang/kite/result/EnumResultHandlerTest.kt
@@ -1,7 +1,8 @@
 package com.tang.kite.result
 
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
 /**
  * @author Tang