+++ kite/src/test/kotlin/com/tang/kite/result/primitive/StringResultHandlerTest.kt
@@ -1,7 +1,8 @@
 package com.tang.kite.result.primitive
 
-import org.junit.jupiter.api.Assertions.*
 import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
 /**
  * @author Tang