+++ kite/src/test/kotlin/com/tang/kite/result/time/DateResultHandlerTest.kt
@@ -1,7 +1,5 @@
 package com.tang.kite.result.time
 
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
 import java.sql.Time
 import java.sql.Timestamp
 import java.time.LocalDate
@@ -10,6 +8,10 @@ import java.time.LocalTime
 import java.time.ZoneId
 import java.sql.Date as SqlDate
 import java.util.Date
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNotNull
+import kotlin.test.assertNull
 
 /**
  * author Tang