+++ kite/src/test/kotlin/com/tang/kite/result/time/LocalTimeResultHandlerTest.kt
@@ -1,13 +1,13 @@
 package com.tang.kite.result.time
 
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNull
-import org.junit.jupiter.api.Test
 import java.sql.Date
 import java.sql.Time
 import java.sql.Timestamp
 import java.time.LocalDateTime
 import java.time.LocalTime
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
 class LocalTimeResultHandlerTest {