+++ kite/src/test/kotlin/com/tang/kite/result/time/SqlDateResultHandlerTest.kt
@@ -1,7 +1,5 @@
 package com.tang.kite.result.time
 
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
 import java.sql.Date
 import java.sql.Time
 import java.sql.Timestamp
@@ -9,6 +7,9 @@ import java.time.LocalDate
 import java.time.LocalTime
 import java.time.LocalDateTime
 import java.time.ZoneId
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
 /**
  * @author Tang