+++ kite/src/test/kotlin/com/tang/kite/result/time/TimeResultHandlerTest.kt
@@ -1,10 +1,11 @@
 package com.tang.kite.result.time
 
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
 import java.sql.Time
 import java.time.LocalTime
 import java.time.LocalDateTime
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
 /**
  * author Tang