+++ kite/src/test/kotlin/com/tang/kite/session/InsertTest.kt
@@ -6,11 +6,11 @@ import com.tang.kite.session.entity.Role
 import com.tang.kite.session.mapper.AccountJavaMapper
 import com.tang.kite.session.mapper.AccountMapper
 import com.tang.kite.session.mapper.RoleMapper
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNotNull
-import org.junit.jupiter.api.Test
 import java.math.BigDecimal
 import java.time.LocalDateTime
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNotNull
 
 /**
  * @author Tang